[Refactor, Style] CardDetailModal: 댓글창 디자인 수정, ProfileIcon: 컴포넌트명 오타 수정, 이미지 비율 일그러짐 fix
